+++ src/lexer/index.ts
@@ -3,3 +3,4 @@ export * from "./utils";
 export * from "./builder";
 export * from "./model";
 export * from "./lexer";
+export * from "./error";